refactor(extension): rename Label -> tools/label
