+++ packages/extension/src/index.ts
@@ -17,7 +17,7 @@ export * from './materials/rich-text-editor'
 export * from './materials/group'
 export * from './materials/node-selection'
 export * from './NodeResize'
-export * from './Label'
+export * from './tools/label'
 export * from './tools/flow-path'
 export * from './tools/auto-layout'
 export * from './bpmn-adapter/xml2json'
