+++ packages/extension/src/tools/label/algorithm.ts
